refactor: moved sharable test utility functions to the `tests` module
